RSpec: split up magick_spec.rb
This test file had a variety of things in it, some of which I left in
place. I removed the `enumerators` spec, as it was testing an internal
detail which I think we should find a way to test via external APIs. I
moved the `_tmpnam_` bit over to the test file that depends on it.
That's also testing private API, so we should refactor that at some
point as well.
