+++ spec/rmagick/class_methods/init_formats_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe Magick, '.init_formats' do
+  it 'works' do
+    expect(Magick.init_formats).to be_instance_of(Hash)
+  end
+end
