+++ spec/rmagick/class_methods/set_log_format_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe Magick, '.set_log_format' do
+  it 'works' do
+    expect { Magick.set_log_format('format %d%e%f') }.not_to raise_error
+  end
+end
